requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
Django==3.2.6
django-common-helpers==0.9.2
django-cron==0.5.1
django-extensions==3.1.3
django-redis==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
docx2txt==0.8
exponent-server-sdk==2.0.0
fuzzywuzzy==0.18.0
idna==3.2
joblib==1.0.1
nltk==3.6.2
numpy==1.21.2
paramiko==2.7.2
pdfminer==20191125
psycopg2==2.9.1
pycparser==2.20
pycryptodome==3.10.1
PyNaCl==1.4.0
pytz==2021.1
redis==3.5.3
regex==2021.8.28
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
tqdm==4.62.2
urllib3==1.26.6
